split no-duplicate-custom-idents into four focused rules
Replace the single combined rule with four independent rules,
one per custom identifier type:

- no-duplicate-keyframes (@keyframes names)
- no-duplicate-container-names (container-name / container shorthand)
- no-duplicate-anchor-names (anchor-name values)
- no-duplicate-registered-properties (@property registrations)

Each rule is self-contained with its own implementation, test suite,
and README. All four are included in the correctness and recommended
configs.
