+++ src/configs/correctness.ts
@@ -5,7 +5,10 @@ export default {
 		'projectwallace/no-unknown-container-names': true,
 		'projectwallace/no-unknown-custom-properties': true,
 		'projectwallace/no-unreachable-media-conditions': true,
-		'projectwallace/no-duplicate-custom-idents': true,
+		'projectwallace/no-duplicate-anchor-names': true,
+		'projectwallace/no-duplicate-container-names': true,
+		'projectwallace/no-duplicate-keyframes': true,
+		'projectwallace/no-duplicate-registered-properties': true,
 		'projectwallace/no-empty-rules': true,
 		'projectwallace/no-important-in-keyframes': true,
 		'projectwallace/no-useless-custom-property-assignment': true,
