+++ src/configs/recommended.ts
@@ -37,8 +37,11 @@ export default {
 		'projectwallace/no-unused-keyframes': true,
 		'projectwallace/no-unused-layers': true,
 		'projectwallace/no-useless-custom-property-assignment': true,
-		'projectwallace/no-duplicate-custom-idents': true,
+		'projectwallace/no-duplicate-anchor-names': true,
+		'projectwallace/no-duplicate-container-names': true,
 		'projectwallace/no-duplicate-data-urls': true,
+		'projectwallace/no-duplicate-keyframes': true,
+		'projectwallace/no-duplicate-registered-properties': true,
 		'projectwallace/no-empty-rules': true,
 		'projectwallace/no-important-in-keyframes': true,
 		'projectwallace/no-invalid-z-index': true,
