+++ src/index.test.ts
@@ -47,8 +47,11 @@ test('exports an array of stylelint rules', () => {
 		'projectwallace/min-selector-uniqueness-ratio',
 		'projectwallace/no-anonymous-layers',
 		'projectwallace/no-atrule-browserhacks',
-		'projectwallace/no-duplicate-custom-idents',
+		'projectwallace/no-duplicate-anchor-names',
+		'projectwallace/no-duplicate-container-names',
 		'projectwallace/no-duplicate-data-urls',
+		'projectwallace/no-duplicate-keyframes',
+		'projectwallace/no-duplicate-registered-properties',
 		'projectwallace/no-empty-rules',
 		'projectwallace/no-important-in-keyframes',
 		'projectwallace/no-invalid-z-index',
