+++ src/index.ts
@@ -38,8 +38,11 @@ import min_declaration_uniqueness_ratio from './rules/min-declaration-uniqueness
 import min_selector_uniqueness_ratio from './rules/min-selector-uniqueness-ratio/index.js'
 import no_anonymous_layers from './rules/no-anonymous-layers/index.js'
 import no_atrule_browserhacks from './rules/no-atrule-browserhacks/index.js'
-import no_duplicate_custom_idents from './rules/no-duplicate-custom-idents/index.js'
+import no_duplicate_anchor_names from './rules/no-duplicate-anchor-names/index.js'
+import no_duplicate_container_names from './rules/no-duplicate-container-names/index.js'
 import no_duplicate_data_urls from './rules/no-duplicate-data-urls/index.js'
+import no_duplicate_keyframes from './rules/no-duplicate-keyframes/index.js'
+import no_duplicate_registered_properties from './rules/no-duplicate-registered-properties/index.js'
 import no_empty_rules from './rules/no-empty-rules/index.js'
 import no_important_in_keyframes from './rules/no-important-in-keyframes/index.js'
 import no_invalid_z_index from './rules/no-invalid-z-index/index.js'
@@ -103,8 +106,11 @@ const plugins: stylelint.Plugin[] = [
 	min_selector_uniqueness_ratio,
 	no_anonymous_layers,
 	no_atrule_browserhacks,
-	no_duplicate_custom_idents,
+	no_duplicate_anchor_names,
+	no_duplicate_container_names,
 	no_duplicate_data_urls,
+	no_duplicate_keyframes,
+	no_duplicate_registered_properties,
 	no_empty_rules,
 	no_important_in_keyframes,
 	no_invalid_z_index,
